refactor(ZD70P1): extract shared manifest-block + dependency-section helpers
Audit (/audit jscpd) flagged the new go.mod/go.work block parsers as a clone, and
the DEPENDENCY_SECTIONS loop + constant were duplicated across fingerprint and
monorepo. Extract:
- manifest-block.ts `readDelimitedBlock(lines, opener)` — the shared `keyword (…)`
  block reader, used by go.mod require and go.work use parsing.
- manifest-dependencies.ts `dependencySectionNames(manifest)` — the shared
  package.json dependency-key collector, used by the fingerprint and the monorepo
  edge model (kills the duplicated section list too).

Behavior-preserving: 43 unit tests + 14 architecture BDD scenarios green, jscpd 0
clones, depcruise 0 violations (158 modules), dogfood --check current.
